fix: install initial aggregate public key to neon, not nakamoto
